moved retry method into Utils class
