+++ src/test/java/iaudiab/maven/dependencytrack/UtilsTest.java
@@ -1,4 +1,4 @@
-package iaudiab.maven.dependencytrack.client;
+package iaudiab.maven.dependencytrack;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -13,9 +13,9 @@ import org.apache.maven.plugin.logging.Log;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import iabudiab.maven.plugins.dependencytrack.client.DTrackClient;
+import iabudiab.maven.plugins.dependencytrack.Utils;
 
-public class DTrackClientTest {
+public class UtilsTest {
 
 
 	@Test
@@ -23,12 +23,11 @@ public class DTrackClientTest {
 
 		{
 			Log log = Mockito.mock(Log.class);
-			DTrackClient client = new DTrackClient("http://localhost", "abc", log);
 
 			Supplier<Integer> nullSupplier = () -> {
 				return null;
 			};
-			assertThrows(CompletionException.class, () -> client.retry(nullSupplier, result -> result == null, 2, 0, 2).join());
+			assertThrows(CompletionException.class, () -> Utils.retry(nullSupplier, result -> result == null, 2, 0, 2, log).join());
 			Mockito.verify(log, Mockito.times(1)).info(Mockito.eq("retry condition met, so retrying after '2' seconds (current retry count: '0'; max. retries: '2')"));
 			Mockito.verify(log, Mockito.times(1)).info(Mockito.eq("retry condition met, so retrying after '2' seconds (current retry count: '1'; max. retries: '2')"));
 			Mockito.verify(log, Mockito.times(1)).info(Mockito.eq("retry condition met, so retrying after '2' seconds (current retry count: '2'; max. retries: '2')"));
@@ -37,28 +36,25 @@ public class DTrackClientTest {
 
 		{
 			Log log = Mockito.mock(Log.class);
-			DTrackClient client = new DTrackClient("http://localhost", "abc", log);
 			Supplier<Integer> valueSupplier = () -> {
 				return 43;
 			};
 
-			assertEquals(43, client.retry(valueSupplier, result -> result == null, 2, 0, 2).join());
-			// "Using api v1 ... " from DTrackClient creation
-			Mockito.verify(log, Mockito.times(1)).info(Mockito.anyString());
+			assertEquals(43, Utils.retry(valueSupplier, result -> result == null, 2, 0, 2, log).join());
+			Mockito.verify(log, Mockito.times(0)).info(Mockito.anyString());
 			Mockito.verify(log, Mockito.times(0)).warn(Mockito.anyString());
 		}
 
 		{
 			Log log = Mockito.mock(Log.class);
-			DTrackClient client = new DTrackClient("http://localhost", "abc", log);
 
 			ConcurrentLinkedQueue<Integer> values = new ConcurrentLinkedQueue<>(Arrays.asList(0, 1, 2, 3, 4, 5));
 			Supplier<Integer> valueSupplier = () -> {
 				return values.remove();
 			};
 
-			assertEquals(3, client.retry(valueSupplier, value -> value < 3, 1, 0, 5).join());
-			Mockito.verify(log, Mockito.times(4)).info(Mockito.anyString());
+			assertEquals(3, Utils.retry(valueSupplier, value -> value < 3, 1, 0, 5, log).join());
+			Mockito.verify(log, Mockito.times(3)).info(Mockito.anyString());
 			Mockito.verify(log, Mockito.times(1)).info(Mockito.eq("retry condition met, so retrying after '1' seconds (current retry count: '0'; max. retries: '5')"));
 			Mockito.verify(log, Mockito.times(1)).info(Mockito.eq("retry condition met, so retrying after '1' seconds (current retry count: '1'; max. retries: '5')"));
 			Mockito.verify(log, Mockito.times(1)).info(Mockito.eq("retry condition met, so retrying after '1' seconds (current retry count: '2'; max. retries: '5')"));
